update code notification <3

+++ LMS library/Controllers/CourseController.cs
@@ -1,8 +1,10 @@
-﻿using LMS_library.Repositories;
+﻿using LMS_library.Data;
+using LMS_library.Repositories;
 using MailKit.Search;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
+using System.Security.Claims;
 
 namespace LMS_library.Controllers
 {
@@ -12,11 +14,15 @@ namespace LMS_library.Controllers
     public class CourseController : ControllerBase
     {
         private readonly ICourseRepository _repository;
+        private readonly INotificationRepository _notificationRepository;
         private readonly DataDBContex _contex;
-        public CourseController(ICourseRepository repository, DataDBContex contex)
+        private readonly IHttpContextAccessor _httpContextAccessor;
+        public CourseController(IHttpContextAccessor httpContextAccessor, INotificationRepository notificationRepository, ICourseRepository repository, DataDBContex contex)
         {
             _repository = repository;
             _contex = contex;
+            _notificationRepository = notificationRepository;
+            _httpContextAccessor = httpContextAccessor;
         }
         [Authorize(Roles = "Admin,Leader")]
         [HttpPost("add-course")]
@@ -28,6 +34,7 @@ namespace LMS_library.Controllers
                 {
                     return BadRequest("Course already exists .");
                 }
+                await _notificationRepository.AddNotification($"{model.courseName} create successfully at {DateTime.Now.ToLocalTime}", Int32.Parse(UserInfo()), false);
                 var newCourse = await _repository.AddCourseAsync(model);
                 return Ok(newCourse);
             }
@@ -79,6 +86,9 @@ namespace LMS_library.Controllers
 
             try
             {
+                var course = await _contex.Courses!.FindAsync(id);
+                if (course == null) { return BadRequest(); }
+                await _notificationRepository.AddNotification($"{course.courseName} deleted at {DateTime.Now.ToLocalTime}", Int32.Parse(UserInfo()), false);
                 await _repository.DeleteCourseAsync(id);
                 return Ok("Delete Success !");
 
@@ -91,7 +101,8 @@ namespace LMS_library.Controllers
         {
             try
             {
-                if (model.id != id)
+                var course = await _contex.Courses!.FindAsync(id);
+                if (model.id != id || course == null)
                 {
                     return NotFound();
                 }
@@ -100,6 +111,7 @@ namespace LMS_library.Controllers
                 var role = await _contex.Roles.FirstOrDefaultAsync(r => r.id == teacher.roleId);
                 if (teacher == null) { return NotFound(); }
                 if(role == null || role.name != "Teacher" ) { return BadRequest(); }
+                await _notificationRepository.AddNotification($"{course.courseName} updated at {DateTime.Now.ToLocalTime}", Int32.Parse(UserInfo()), false);
                 await _repository.UpdateCourseAsync(id, model);
                 return Ok("Update Successfully");
             }
@@ -132,6 +144,15 @@ namespace LMS_library.Controllers
             {
                 return BadRequest();
             }
+        }
+
+
+
+
+        public string UserInfo()
+        {
+            var result = _httpContextAccessor!.HttpContext!.User.FindFirstValue(ClaimTypes.NameIdentifier);
+            return result;
         }
     }
 }
